up to v0.12.2

+++ dl-checksum.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env sh
-VER=v0.12.0
+VER=v0.12.2
 DIR=~/Downloads
 MIRROR=https://github.com/open-policy-agent/opa/releases/download/$VER
 
@@ -11,10 +11,15 @@ dl()
     FILE=opa_${OS}_${PLATFORM}${SUFFIX}
     LFILE=$DIR/opa_${OS}_${PLATFORM}_${VER}${SUFFIX}
     URL=$MIRROR/$FILE
-    wget -q -O $LFILE $URL
-    printf "# %s\n%s_%s: sha256:%s\n" $URL $OS $PLATFORM `sha256sum $LFILE | awk '{print $1}'`
+    if [ ! -e $LFILE ]
+    then
+           wget -q -O $LFILE $URL
+    fi
+    printf "    # %s\n" $URL
+    printf "    %s_%s: sha256:%s\n" $OS $PLATFORM `sha256sum $LFILE | awk '{print $1}'`
 }
 
+printf "  %s:\n" $VER
 dl linux amd64
 dl darwin amd64
 dl windows amd64 .exe
